agents/requirements.txt
google-adk==0.4.0
python-dotenv==1.0.0
requests==2.32.0
